requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
email-validator==2.0.0
openpyxl==3.1.2
pywin32==306
requests==2.31.0 